Record implicit modules in the modules table
Summary:
This diff fixes yet another UGE incremental update bug, where a previously
nonexistant implicit module coming into existence would fail to trigger a recheck
of code that had an unresolved import error related to that package.

In this case, the bug was caused by us bypassing the `Modules` table for
implicit modules, and instead directly creating implicit module metadata on the
fly (and hence without dependency tracking!) inside UGE ReadOnly functions.

Instead, in this diff we push the creation of implicit module metadata down
into the lazy loader and we actually store it in our module table. All reads
then go through the lazy loader, and this will cause dependencies to be tracked
and triggered in the normal way.

Note that in addition to the new implicit module metadata tests designed
to catch this bug, we also catch a bug where one of the `resolve_exports` tests
was verifying incorrect behavior by missing an implicit module load that
happens as part of that logic.

NOTE: we still have a bug on implicit module *removal* not being propagated.
I'll work on that next, but in practice this is less of a problem; it is *very*
common for someone to have a missing module and then add it (via a new buck
dependency or pip install) so the missing-to-not-missing case is a major UX
problem. It is less common to accidentally remove a dependency. As a result, I
think we should go ahead and ship the incomplete fix.

Reviewed By: pradeep90

Differential Revision: D45091842

fbshipit-source-id: a201ffabd5a5e4862b0fa1dbf0a0846912ffd666