Allow-list staticmethod and classmethod for decorated override tests
Summary:
Looking at pyre-binary signals for D45005695, it feels like we can do a little better - we're suppressing errors for some obvious override mistakes that are decorated with `staticmethod`.

We should probably handle decorators in a more principled way in the future, but for now it seems reasonable to allow a few ones that are known to not modify the function signature for override-checking.

I don't want to hold up D45010870 so I'm publishing this separately.

Reviewed By: kinto0

Differential Revision: D45058619

fbshipit-source-id: b2bf31ca1275152679785e2d35585a8443deb87c