Rename load_module_if_tracked to try_load_module
Summary:
At this point we attempt a load no matter what; if the module is
untracked we wind up storing a `None` in the ModuleMetadata map,
which is necessary to make dependency tracking work correctly if
modules pop in and out of existence.

So, let's call it `try_load_module`; this is also consistent
with the related function `try_load_module_with_cache`, which sort
of does the same thing but short-circuits for already-cached modules.

Differential Revision: D45119331

fbshipit-source-id: 5bd23c9ef4a2f0bc3036074b2b49b5750c8f8ed7