Add more inline docs on lazy loading + change function name
Summary:
This is in response to comments in diff review that the semantics weren't
clear. I'm honestly not sure we can make them super clear, because the
semantics are coupled with qualification logic which is actually a bug.

But I think a comment with an example of a fully qualified name, and why
we cannot know what module that name is in, helps explain both why we
are loading ancestors and why the name is "load 'all possible' modules":
we're loading all of the modules where the name we actually want to look
up might possibly be defined.

I also changed `load_all_possible_modules` to `load_all_possible_modules_for_symbol` in an attempt to better approximate in just a name what it's really doing.

Differential Revision: D45119332

fbshipit-source-id: f2af91a535c18b3543b45ef0536e3f4267e61da1