Add position information to taint.config errors
Add position information to taint.config validation errors -
RuleCodeDuplicate, SourceDuplicate, SinkDuplicate.

Add an additional option to the analyze command called
--verify-taint-configuration which just verifies the taint configuration
and nothing more.

What makes this commit long and sturdy is mostly due to YoJSON, the
current json file parser used to parse taint.config files, lack support
for line numbers and column numbers of matched elements. This leaves us
with two options:
1. Use another json parser which has support for getting line numbers
2. Get positioning information via hacky string operations

2. Has performance implications and for 1, the closest parser that does
this is Jsonm, which has a complete lack of utility functions that
yojson provide.

I opted for 1 and then converted the results to conform to the yojson
data format, so that parsing is done only once and we get to use
yojson's utility functions along with positioning information!

Why is this important? For the visual code extension, positioning
information is vital.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>