Allow global_leaks query to take in list of callables to analyze
Summary:
**Background**: This diff enables the global leaks query to take in a list of callables to analyze in addition to the single callable it takes on now.

```
pyre query "global_leaks(my.callable)" # works already
pyre query "global_leaks(my.callable_1, my.callable_2, ...)" # works now
```

**Context**: We would like to eventually have the opportunity to parallelize the analysis of each function during a global leak check. With the current set up, we're passing in each callable we'd like to analyze in a batch query, so there are three approaches we can take for this:
1. Parallelize all batch queries.
This is probably not a great approach, since it's likely that not all queries can be parallelized (or it could become a problem in the future).
2. Add a special case to the `batch(...)` query to collect all callables to be analyzed in nested `global_leaks(...)` queries, and run those in parallel.
This approach also doesn't seem like a great approach, since special casing the `batch` query to handle `global_leaks` differently seems like bad design and will lead to code that is more difficult to maintain.
3. Pass in all callables we want to analyze to `global_leaks(...)` in a list, and parallelize just that operation.
This approach seems the best to me, since it allows the `global_leaks(...)` query to still cleanly consume its parameters in a way that should makes sense to clients. It also allows us to easily add more parameters to global leaks in the future if we see a need for it.

The followup diff will add a new datatype to better represent the query responses.

Reviewed By: pradeep90

Differential Revision: D44959919

fbshipit-source-id: 1128304e091856eed0564a1e6accd8e1a2acd46f