Return errors and global leaks directly from global_leaks query instead of batch
Summary:
**Background**: This diff adds a new response type and data structure to represent what is being returned from the global leaks query. Instead of operating in a batch query and returning a response of lists of query `Error` and list of global leak `Single (AnalysisError.Instantiated.t)` (which requires extra work to parse and classify on the client side), a data type to better represent the return values is provided instead.

> Note: this will cause a breaking change to the API that the follow up diff will fix.

**Context**: We would like to eventually have the opportunity to parallelize the analysis of each function during a global leak check. With the current set up, we're passing in each callable we'd like to analyze in a batch query, so there are three approaches we can take for this:
1. Parallelize all batch queries.
This is probably not a great approach, since it's likely that not all queries can be parallelized (or it could become a problem in the future).
2. Add a special case to the `batch(...)` query to collect all callables to be analyzed in nested `global_leaks(...)` queries, and run those in parallel.
This approach also doesn't seem like a great approach, since special casing the `batch` query to handle `global_leaks` differently seems like bad design and will lead to code that is more difficult to maintain.
3. Pass in all callables we want to analyze to `global_leaks(...)` in a list, and parallelize just that operation.
This approach seems the best to me, since it allows the `global_leaks(...)` query to still cleanly consume its parameters in a way that should makes sense to clients. It also allows us to easily add more parameters to global leaks in the future if we see a need for it.

The previous diff adds the changes required to take in a list of callables to analyze. The followup diff will fix the breaking API change.

Reviewed By: pradeep90

Differential Revision: D44959915

fbshipit-source-id: d84570b06d8efbc2374d8ed887bad1417350fc34