Duplicate endpoints to prevent global leak check from breaking with API changes
Summary: **Background**: In upcoming diffs, the API request and response for the global leak check query will be changed. To prevent confusion between people developing against this query, a new `global_leaks_DEPRECATED` endpoint is added, which copies the current implementation of the `global_leaks` query, to ensure users of the analyze_leaks.py script won't be presented with API change issues.

Reviewed By: pradeep90

Differential Revision: D45057468

fbshipit-source-id: 981db5e9ca11d5b2912f2db8fc193138a2155c15