Revamp how we calculate suppressed errors and unused ignores.
Summary:
Problem: We were using a side-effectful approach before, using two hash tables and one mutable variable that was set within a nested function that was called in a `List.iter`.

just_why

Switching to a much more straightforward approach:
+ For each error, get the pyre-ignores that cover it. That allows us to get the un-suppressed errors.
+ Flip that to get the errors that are suppressed by each ignore. Go over all ignores to find ignores that are do not suppress errors at all or, in case of multiple ignore codes, only suppress some of their ignored codes.

I don't think this should be much costlier. But RFC if people feel strongly that we need to be super-efficient here and avoid list processing.

NOTE: We get spurious unused-ignore errors for multi-line f-strings. Coming diffs fix that.

Reviewed By: grievejia

Differential Revision: D45096970

fbshipit-source-id: 076488a627a4dad94cf0e89a39ec5b03ef6e0e12