If ModuleTracker invariant check fails, give a useful error
Summary:
We spent some time working out a bug today that turned out to be caused
by a violation of the invariant check in ModuleTracker (the root cause was two
pyre servers using the same link tree, which made logic depend on
race-condition-sensitive filesystem operations)

It took longer to start really debugging because we couldn't see where the problem
was (it was ultimately the `is_external` field). Having a useful error message
that include the actual mismatched module paths would have saved us some time.

Reviewed By: grievejia

Differential Revision: D45251467

fbshipit-source-id: 0f7216854f084780ed5e8735be3b2d0f21b6a0ad