Add a test demonstrating an imprecise call graph with inner functions and decorators
Summary:
As titled.

For context, we use a set of heuristics to bypass some decorators (including `lru_cache`). Those are hardcoded in the analysis itself. Unfortunately, this does not work well for all cases (as shown here).

In the future, I would want to get rid of most of these and use proper models (such as `IgnoreDecorator`) instead.

Reviewed By: tianhan0

Differential Revision: D45319575

fbshipit-source-id: a83634b03a03169fd46610a9c78da8bec2830e81