Ignore functools decorators
Summary:
We previously had special logic in the call graph construction to find the right edge when calling a function decorated with `lru_cache`.
However, that logic fails on some corner cases, see D45319575.
At this point, I believe it is easier to ignore entirely some specific iterators in the context of pysa.
This diff does exactly that for the most common decorators in `functools`.

Reviewed By: alexkassil

Differential Revision: D45321457

fbshipit-source-id: 9837ab2ab2299df9d991c8e127614a5fee4dbedf