Remove global leak 3100
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff removes the old global leak error with generic error code 3100.

Reviewed By: pradeep90

Differential Revision: D45300276

fbshipit-source-id: 6c86cdd1499dc076968d003d83c506a69434fdac