Codify globals written to locals
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds `WriteToLocalVariable` for globals that are written to locals. If the number of such occurences are very high, it might be worthwhile to do a local taint analysis to flag only those locals that are modified.

Reviewed By: pradeep90

Differential Revision: D45300190

fbshipit-source-id: 86e0fdba54a2609f4d6f1bf4f470606ab747e1ab