Codify writes to class attributes via settr functions
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds `WriteToClassAttribute` for class attributes that are modified using `setattr` functions. Currently it does not identify the type of the class attribute that is being modified. This will be investigated and fixed in an upcoming diff.

Reviewed By: pradeep90

Differential Revision: D45046585

fbshipit-source-id: 6e96229bcda0471ef1cd48e04a592e6b9bb707a2