Codify globals passed as method arguments
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds `WriteToMethodArgument` for globals that are passed as function arguments.

Reviewed By: pradeep90

Differential Revision: D45300159

fbshipit-source-id: fb4dfbce2f04bcdb2068f8f45003b5f12728857c