Add an interface for calling a remote code index
Summary:
For find references to work with Glean, we need to make Glass specific calls, which OSS version of Pyre would not need.

To accomodate, I'm adding an interface for a remote code index that will provide info based on the current file path and the position.

The next diff will add calls to Glass to fetch back references.

Reviewed By: grievejia

Differential Revision: D44017478

fbshipit-source-id: b15c47cb60ab4777cecb8b14e41e1cda8483a4db