Do not cache the override graph
Summary:
Unfortunately, the override graph depends on the taint configuration and on the set of models (because of `SkipOverrides` annotations). Therefore, it cannot be cached safely, since we assume the cached state is unchanged when models change.

For the same reason, the call graph is not currently cached.

In the future, we could probably be smarter and only invalidate part of the analysis based on what changed.

Reviewed By: alexkassil

Differential Revision: D45395428

fbshipit-source-id: 2aeee91c27e57dc0b109ea7c58a548aed8f6909f