Remove subkinds from model functions tainting dataclass fields
Summary:
Thanks to the new pysa capability to surface parameter name in source and sink names (D44725465) we do not need anymore to have an explicit subkind.
This diff removes `taint_callable_with_parameters_names` which was used to taint all the parameters adding the subkind information and modifies `taint_callable_dataclass_fields_parameters` to create models both for  dataclasses fields and simple parameters.

Plan is to land
1.  Land configerator change to disable old model generator `get_ig_django_view_sources_with_parameter_and_dataclass_fields_as_subkind` in lineage runs D45179758
2. Land Instagram change removing  `get_ig_django_view_sources_with_parameter_and_dataclass_fields_as_subkind` and creating `get_ig_django_view_sources_dataclass_fields_precision` D45179511
3. Land Configerator change to enabled this new mode `get_ig_django_view_sources_dataclass_fields_precision` D45179992
4. Land FBcode diff to remove subkinds D45185969
, D45180363  (this diff)

Reviewed By: arthaud

Differential Revision: D45185969

fbshipit-source-id: 737f21126cd7fece630a24836c039c57d0777963