Created function extract_parameters_with_types
Summary: In the diff on top we need to extract all the parameters of a callable both annotated and not along with the type of the parameter. For this reason I created `extract_parameters_with_types` and moved the functions needed from the function_tainter file to the more approapriate inspect_parser.

Reviewed By: arthaud

Differential Revision: D45180363

fbshipit-source-id: 078ae695e7b6d7722c801ede1b1413bbfd4cd5e1