Add the type of class attribute mutation in assignment target
Summary:
Context: We want to group all writes to mutable datastructures for analysis. Currently the error message for write to class attribute does not give this level of detail. It just provides the class which has this error.

This diff implements the change to print the actual attribute and its type for assignment target. An upcoming diff does the same change for arbitrary expressions.

Reviewed By: connernilsen

Differential Revision: D45446313

fbshipit-source-id: 069c834f18426ca963c651cdad4d619e3f4d7699