Implement json parser and migrate taintConfiguration
Implements a json parsing module inside jsonParsing.ml that adds support
for line and column numbers based on Jsonm. Removes all dependencies on
YoJSON which doesn't have any support for positions.

Since, Jsonm is pretty bare bone, implemented some utility functions
inside a Util module that does pretty much what YoJSON's utility
functions did.

There are minor changes in taintConfiguration to accomodate for the
change.

This is a part of commits that seek to remove dependency completely on
YoJSON and switch to JSONM because YoJSON doesn't support line numbers
in any way which we require for descriptive error messages.

Modifies and updates confirmationTest. Since we use a proper json
parser, the json files for testing have been updated to use the proper
json format as well.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>