remove aliases to non-existent modules
Summary:
the `Sort` module was removed in 4.08.0 (https://ocaml.org/releases/4.08.0) and modules `Lock` and `Socket` don't seem to have corresponding `.ml` files on disk.

the reason this has gone undetected before now is that i believe dune always defines `-no-alias-deps` (which it shouldn't when it doesn't have to) and either (the dune build) doesn't enable warning 49 (warn on missing `.cmi` files) or doesn't treat warnings as errors (or both or something idk ¯\_(ツ)_/¯).

Reviewed By: grievejia

Differential Revision: D45692664

fbshipit-source-id: a710bb9d041526f9131a135de9152b1122476de2