Add an option to include external sources
Summary:
# Context
Pysa users write DSL queries to annotate callables that produce sources or sinks.
To properly test that a DSL query annotates the right callables, one would have to wait for a full run of `pyre analyze` (i.e, pysa), which can be quite slow.
To improve the user experience, `pyre query "model_query(...)"` allows to run a DSL query and get its results, without performing a full run.

# Problem
There are currently discrepancies between `pyre query model_query` and `pyre analyze`, since the latter does include external sources, i.e files that are in the search path but NOT in the source path. This is because the type checker does not need to type check those files.

If a DSL query matches a function defined in an external source file, it will be found during a `pyre analyze` run but not in a `pyre query model_query()`. This can be quite confusing.

# Solution
To fix this problem, let's add a flag to `pyre start` that enables the analysis of external sources, which will remove the discrepancy.

Reviewed By: stroxler

Differential Revision: D45694783

fbshipit-source-id: 6b051b39208f1e3f0c1707cf35e4857e68594fac