Do not implement sexp on query types
Summary:
The pyre query request and response type were implementing `sexp` in the only purpose of logging within tests.
This is preventing types that don't implement `sexp` (for instance, `Yojson.Safe.t`) to be used in query responses.
This is also very contagious, leading to a lot of types implementing `sexp` for the only purpose of logging.

For the same reasoning, this diff replaces the use of `compare` by `equal` (`Yojson.Safe.t` implements `equal` but not `compare`).

We will use this in a follow-up to store `Yojson.Safe.t` directly in a response type.

Reviewed By: tianhan0

Differential Revision: D45732772

fbshipit-source-id: fa60af12034b42b511ae3f826d53084005367071