Emit errors when successfully matching a literal dictionary against a union of `TypedDict`s.
Summary:
Problem: When matching a literal dictionary against a union of `TypedDict`s `Movie | Book`, we check it against each item in the union to see if it can be given the expected type. If we find one that matches, we just return that without any errors.

However, there can be errors, such as extra fields, which get suppressed.

Solution: Instead of just checking if there exists a successful match, get the match and pass on the typed dictionary errors to the user.

Note: I'm sorting by the number of errors to get the match with the least number of errors.

Reviewed By: grievejia

Differential Revision: D45713894

fbshipit-source-id: fc8d1bad042307ce63d589890d79f3da8a460dda