Add deduplication + deterministic order to `find_module_paths`
Summary:
This diff closes a minor bug illustrated by a new test in the previous commit, where
we would multiply-count a module if the user provided both a containing directory
path *and* a direct path (or multiple directory paths).

We do this by deduplicating at the module path level rather than earlier when paths
can be directories.

I also make the output deterministic by sorting; technically this isn't needed, but
it's slightly easier to debug and test if the ordering is easy to skim, and the
cost of sorting even a large list of strings is negligible compared to the rest
of `pyre report`.

Reviewed By: jasontatton

Differential Revision: D45820278

fbshipit-source-id: ee6a39b229591bfd8dac7667d0e40f9b6eb8c3b6