Add trailing commas to some report-related code.
Summary:
I dislike having multiple parameters on the same line in Python, it's much
easier for me to skim code where parameters are all left-aligned. This isn't
a style convention I try to force on other people, but for code where I'm
the main maintainer it makes a difference.

I was skimming the report code today with Jia to see how the path filtering
works and was irritated by a few of these functions - adding a trailing comma
causes Black to format them the way I find easier to read.

Reviewed By: jasontatton

Differential Revision: D45818674

fbshipit-source-id: 7f73c4bd51043a3e07dee2685dde41e7e4e0aacb