Remove get_paths_to_collect
Summary:
As discussed in an earlier commit, this function doesn't do
very much and the code is easier to follow if we move the one
really useful thing it does (falling back to [root] when no paths
are provided) into callers.

This commit deletes the function and inlines the fallback. In the process
I also flatten out some of the control flow for `expression_level_coverage`
(which is more complex than the report / statistics logic because it allows
passing an argument path, which)

Reviewed By: jasontatton

Differential Revision: D45820769

fbshipit-source-id: 9770f1d39f63364298dafad595a4d895de0b2804