Preserve classmethod name if the decorator doesn't change its signature.
Summary:
Problem: When applying a decorator on a function, we preserve the function's name if the type signature is unchanged. This leads to better error messages (i.e., "In call `foo`" instead of "In anonymous call").

However, when we have a classmethod that is decorated, we always lose the function name.

Reasoning for why we preserve the name only when the type signature is preserved: D21487479.

It basically said that we want to preserve the same type for a given global name. Not sure if that restriction holds today, but I'm preserving it anyway.

Reviewed By: connernilsen

Differential Revision: D45801301

fbshipit-source-id: 0f52e2add0cfe03d1ec76e88fe6d6fd9d958b554