non-blocking requests
Summary:
Context: The language server request handler is currently fully synchronous. It blocks on a request being returned from the daemon before even reading the next request from vscode. This is a problem because it:
- results in incorrect logging in our telemetry events because the "start" of the timer is not when the request got to us, but when we started processing it
- can result in incorrect daemon status because we wait until the daemon is ready from a request + potentially restarted before logging anything
- will block any requests that don't require querying the daemon

This diff is a first attempt at making the request handler non-blocking. It uses `asyncio.Task`s to track each nonblocking request. I make the distinction on nonblocking because shutdown requests and exit requests should not send control to a different thread, they should drop everything and shut down immediately.

Some callouts:
- `dispatch_nonblocking_request` currently calls `_try_restart_pyre_daemon` if the daemon is not running. Because `is_task_running()` is checked first, we should not be starting up pyre daemons in excess of one at a time. I would like to keep this logic the way it is so concurrent requests will correctly log uptime (they'll know the daemon is down and not wait on restarting the daemon).

If people are on board with this change, I will implement a background task manager for these similar to `background_tasks` in a future diff to encapsulate this logic better.

Reviewed By: grievejia

Differential Revision: D46072422

fbshipit-source-id: 48f421afeb0b307436bce4f119fe85924d68e2b3