Log codenav Exceptions to Scuba
Summary: I noticed that we log exceptions to scuba in the [typechecking server](https://fburl.com/code/v3mhu8nv) but not the codenav server. In the event of a crash, I would like to see logs other than the vscode chronicle logs. This diff logs on any caught request exceptions.

Differential Revision: D46192862

fbshipit-source-id: e44a6148c011c9969f4f93c45283461dbe790ee6