Calls to nested functions should not be "anonymous".
Summary:
Problem: Nested functions are stored in the local resolution with a callable type. However, we mark them as anonymous, meaning that error messages show `In anonymous call, ...`, which is confusing.

Preserve the callable name.

This also addressed related bugs, which I've captured in tests.

Design choice: The callable name is a local name. I'm keeping it as such instead of converting it to `outer.inner`, because Pysa tests depend on this.

Reviewed By: grievejia

Differential Revision: D45967775

fbshipit-source-id: 2aa0fa9cc0bc81c7fad67759eec05350ffd39123