Allow `ParamSpec` decorators to be compatible with decorator factory protocols.
Summary:
## Context

When checking callable `f` <: callable `g`, we "call" `f` with the parameter types of `g`. If `f` is generic, we solve for those types using the types from `g` and instantiate the return type of `f`. We call this `simulate_signature_select`.

For example, in `Callable[[T], list[T]]` <: `Callable [[int], list[int]]`, we call the left callable with argument types `[int]`.

+ That adds constraints `int <: T` to our existing constraints.
+ We partially solve for the free variables of the left callable (in this case, `T`), to get `T = int`.
+ We instantiate the return type of the left callable to get `list[int]`.
+ We return that along with the existing constraints altered with the knowledge that `T=int`. This is usually needed when the right callable is generic.
+ `solve_less_or_equal` will check that the instantiated return type of the left callable (`list[int]`) <: the return type of the right callable (`list[int]`).

When both callables are generic, such as `Callable[[T1], T1] <: Callable[[T2], T2]`, calling the left callable with `[T2]`, leads to two constraints:
+ `T1` has lower bound `T2`
+ `T2` has upper bound `T1`.

We partially solve for `T1` to get `T1 = T2` and return the information that there existed a solution. That seems pretty straightforward.

## Problem

When we have `Callable[P1, R1] <: Callable[P2, R2]`, we just add **one** constraint instead of two: `P2` has upper bound `P1`.

Now, when we try to solve for the left callable's free variable P1, we don't see it in the set of constraints. (Only `P2` has any bounds.) So, we return "no solutions found" :'

This leads to confusing errors such as those in T128671942.

## Solution

Add bounds for both `P1` and `P2`, just like we do for `T1` and `T2`.

## Followups

Follow-up diffs: Will document the code with all of this.

Reviewed By: jasontatton

Differential Revision: D46244753

fbshipit-source-id: ef7616696b6027954556aa49421c6c549b00947d