Adds verify taint config only option
Adds verify_taint_config_only option to the analyze command that just
verifies taint.config files and skips the analysis. This can be useful
in the future when we want to verify taint config files via the vs code
extension without performing analysis or even any other sort of
preprocessing.

Modifies the ocaml server and the python client to pass options for the
same.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>