Remove the state argument from the map in map_reduce
Summary:
The `map_reduce` interface currently takes a `map: 'state -> 'input list -> 'intermediate` and a `reduce: 'intermediate -> 'state -> 'state`. It is a bit curious that the `map` takes a state. In practice, we almost always ignore the `state` in `map`. This is only correct because the implementation always passes the initial state to `map`.

In conclusion, we might as well remove the `state` passed in to `map` to prevent confusion. If the initial state is required, it is usually straightforward to provide it.

Reviewed By: pradeep90

Differential Revision: D46289805

fbshipit-source-id: dccad177af1fd27505e983bb223ede2b3b84d4fd