Add logging to pyre code nav client startup
Summary:
When pyre's code navigation client starts up, it waits on a request from an IDE. However, it is confusing when run from the command line since the last log seen is

```
INFO Could not determine the number of Pyre workers from configuration. Auto-set the value to 35.
```

This diffs adds a local log before the client starts waiting on a initialization request. This diff also updates some basic descriptions.

Reviewed By: kinto0

Differential Revision: D46359735

fbshipit-source-id: af3ff77f522dd6c29e5e12ab859ab4fdf13927f5