Display positions in SourceDuplicate errors
Previously when SourceDuplicate errors occured, we just left a note
specifying the duplicate name. Now we can specify the locations where
the duplicates occured.

Modifies annotationParser as annotationParser.source_or_sink is where
the sink or source is parsed into from taint.config files to contain the
location information to be used incase duplicates occur.

Modifies and updates tests to confirm and check the new record field and
type of SourceDuplicate error.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>