Add pyre status box in VSCode for code navigation
Summary:
We display a status message in the bottom of your IDE for pyre's typechecking language server status. This is a LSP message from the pyre language server to the vscode extension language client. The goal of this diff is to turn this on for code navigation.

We also update relevant test cases as we now forward CodeNav events to the window

Differential Revision: D46394130

fbshipit-source-id: c6849c778339e399fe36432555ccbe416253bc84