Use --save-results-to when running run_cache_test.py
Summary:
Such that later we can read from `taint-metadata.json` about how different
parts of the cache are used in a pysa run. Such information will be used to
test the behavior of cache. This is an ideal way to test the cache, because it
directly tests the usage of the cache. By contrast, currently, to test cache we
compare the analysis results, which is an indirect and therefore less ideal
way.

Reviewed By: arthaud

Differential Revision: D46580005

fbshipit-source-id: f78320032a1c54cc3504b81bba64eda9f91bd99c