Restructure GetDefinitionLocations response
Summary:
As part of T154402456 we need to log whether the responses to the `Go To Definition` request are glean indexed results or results from the pyre client. We serve glean indexed results if the pyre client is not ready (behind GK), implementation here D46574734.

Currently the telemetry logs are written at [pyre_language_server.py](https://fburl.com/code/xi7hf597), but it is only at a lower level of the RemoteIndexBackedQuerier daemon_querier that has context on whether an indexed result can be used.

Thus we either refactor the telemetry logging to occur at a lower level, or restructure the query response to include additional information. This diff chooses the latter approach.

As this decision will affect the refactoring efforts of a lot of the event handlers (if indexed results are to be integrated for other features), I wanted to gather some thoughts on how best to refactor this.

Reviewed By: kinto0

Differential Revision: D46612974

fbshipit-source-id: 0a2fe5e19ed73f2821f26229a03b51086fd0049f