Implement v0 of pyre report-any-expressions
Summary:
This v0 implementation is dead-simple, because it just converts
from the well-typed but not-well-formatted data produced by
`expression_level_coverage` into a different format that is nicer
for hive tables.

I don't think the underlying logic will last very long: it's likely
we're going to want to parallelize the query (which will require backend
changes and then frontend changes).

But I want to get a test CI workflow running as soon as possible, even if it is
slow, so I'd like to ship this and get it into the nightly release tonight
if possible.

Reviewed By: kinto0

Differential Revision: D46653130

fbshipit-source-id: 822f8fcc1abc78b99fcef29f8260fb09077264f9