Allow not stopping the cache test when a subtest fails
Summary:
Currently, it is time-consuming to update the expected results of the cache
tests, because as long as one of the subtest fails, the whole test stops and
therefore the remaining subtests cannot be run.

This diff allows running all subtests, regardless of whether some of them
fails. This will improve the productivity when working on run_cache_test.py.

Reviewed By: alexkassil

Differential Revision: D46607893

fbshipit-source-id: 71cbe4e0d5da7ceb9ef8f8da60872a7891165c5f