Remove unused imports
Summary:
Unfortunately I can't figure out how to get flake8 to show up as
a diff signal (which, if it did, would trigger lintBounce for me).

As a result I get unused imports sometimes - this time it's from
removing the unused error stats logging code yesterday.

Github CI flags it, which is nice.

Reviewed By: inseokhwang

Differential Revision: D46700466

fbshipit-source-id: 648fe63c1a39f6a4419de5ebd77a78b847e70e58