Conditional serve glean definition in RemoteIndexBackedQuerier
Summary:
RemoteIndexBackedQuerier is a querier that has access to both pyre's daemon and Glean.

If the either:
- Pyre daemon status is not READY
- Pyre daemon returns an error result (in this case, we still logged in telemetry)

We return the result from calling the glean API

This increase uptime of our code navigation feature

Note that GK was  fetched directly instead of through the pyre GK utility since the latter case will require users to toggle the `disable GK cache` option which may slow adoption.

Reviewed By: grievejia

Differential Revision: D46574734

fbshipit-source-id: 5fc7b54d2180119bb425cd9cea946cdc967c3f5f