Add and test telemetry extra to distinguish indexed results
Summary:
# Overview
This diff adds a telemetry extra for whether result came from glean or pyre. With the integration of the indexed results we want to see how useful this is in production.

With these results we intend to monitor how often the pyre daemon returns results vs how often it has to resort to Glean

Also added a missing `setup.async_test` header and debugged the associated test

Reviewed By: kinto0

Differential Revision: D46631375

fbshipit-source-id: 3b4ba60c453f28fc321853a6944bc273ea39555c