Create a map from different parts of the cache to their usage status
Summary:
Such that we record the cache usage status in each Pysa run, such as whether
the type environment was loaded from cache, and if not, what was the reason.

An example of recorded cache usage is:
```
{
  "shared_memory_status": {
     "Loaded": {
        "TypeEnvironment": "Used",
     }
  }
}
```
Here:
- `shared_memory_status` is about whether the previously saved shared memory is
  not loaded (i.e., disabled), or loaded and then invalidated (with what
  reason), or loaded and then ready to use in the next pysa run.
- When `shared_memory_status` is `Loaded`, then `TypeEnvironment` is about
  whether different parts of the loaded shared memory is used in a pysa run.

This diff makes the following changes:
1. **Core Analysis.** Record the cache usage of the shared memory and the type
environment, and then write the usage into `taint-metadata.json`.
2. **Cache Tests.** Upate run_cache_test.py to test the cache in two ways.
First, we compare the analysis results, which is the existing behavior. Then,
we compare the usage of cache.

The next diff will record the cache usage of other parts in the shared memory,
such as the class hierarchy graphs.

Reviewed By: arthaud

Differential Revision: D46575017

fbshipit-source-id: 79ab5b6955fd6a70a11572c1be8264e9a05ade5e