Add Pyre client argument for call hierarchy
Summary:
This stack of diffs will focus on integrating `Call Hierarchy` functionality into VSCode. Call hierarchy is another API glass provides as part of the glass searchRelated request. It is very similar to find references but it also shows transitive references. This would be a huge productivity win for developers.

This diff adds the necessary client arguments

Reviewed By: kinto0

Differential Revision: D46730041

fbshipit-source-id: d8ba98e797372e39339bd3854aedb7cdf21bf42e