Temporarily undo logic to conditional serve glean definition in RemoteIndexBackedQuerier
Summary:
Original commit changeset: 5fc7b54d2180

Original Phabricator Diff: D46574734

This diff temporarily reverts the logic to conditionally serve glean definitions while the pyre_daemon is not ready. We observed large latency with the initialization of the glass client which blocks the go-to-definition query for longer than intended.

We are looking to address the initialization delay, in the mean time we revert this new feature

Reviewed By: kinto0

Differential Revision: D46743734

fbshipit-source-id: 06c8dd9abd29330294a84d3cf8416a05c927a53d