Relativize pyre report paths to project root
Summary:
When I wrote the `pyre report` command, I was running pyre report
from the project root. This led me to get mixed up about path
relativization - I relativized to the curent working directory. But
what we actually want is path relative to the project root (i.e. local
root if there is one, otherwise global root), so that we aren't duplicating
information in the `local_root` vs `module_path` portions of the output
datasets.

This diff makes that explicit: compute the project root, and relativize
paths against it when we create output data.

Reviewed By: grievejia

Differential Revision: D46861815

fbshipit-source-id: de910558202cb295b268559041fd2555c30e58e2