Don't emit line and column as `-1:-1` when an operand has a readonly mismatch.
Summary:
Problem: When we match arguments to parameters, we set the expression for `self` as `None`. Also, if we get a type error from an operand, we try inverting. This means that, if there is a type error when checking trying the inverted operand (i.e., `argument.__eq__(self)`), we would end up using the location of the error message as `Location.any`:

https://www.internalfb.com/code/fbsource/[cd4c8b2d1bea]/fbcode/tools/pyre/source/analysis/attributeResolution.ml?lines=1729-1734

The deeper fix (T135849678) is to make `signature_select` pass in the `self` expression:

https://www.internalfb.com/code/fbsource/[cd4c8b2d1bea]/fbcode/tools/pyre/source/analysis/attributeResolution.mli?lines=266-274

For now, I'm unblocking this by falling back to the location of the operand.

Reviewed By: stroxler

Differential Revision: D46911804

fbshipit-source-id: acab83781227512cf35ba7b7801f92efd0653348