Print a warning on slow type checks
Summary:
We already have a mechanism to log the slow analysis of a callable in pysa, using the `Metrics` module.
This diff renames `Metrics` into `Alarm` and moves it into the root directory, so we can use it to warn about slow type checks of callables.
This will be useful in finding performance bottlenecks.

Reviewed By: alexkassil

Differential Revision: D46934260

fbshipit-source-id: 809b97e083e93f689482f51536b7cf7cf67b211c