Reintroduce logic to conditional serve glean definition in RemoteIndexBackedQuerier
Summary:
Original Phabricator Diff: D46574734

Temporarily reverted in Diff: D46743734

This diff reintroduces the logic to conditionally serve glean definitions while the pyre_daemon is not ready.

We observed large latency with the initialization of the glass client which blocks the go-to-definition query for longer than intended, which resulted in us temporarily rolling back the feature. However, this latency is now addressed in D46743734, and so we now reintroduce the feature

Reviewed By: kinto0

Differential Revision: D46913743

fbshipit-source-id: dd082bd34a6304b0051d2380c2abb9ec55887178