Use SharedMemory.FirstClass when implementing ClassIntervalSetGraph.SharedMemory
Summary:
For context, we want to uniformly use SharedMemory.FirstClass to
enable caching the results from various Pysa steps, such as computing the class
interval graphs and the overriding graphs.

This diff also fixes a bug in the integration tests that we did not reset the
shared memory for the class interval graphs after each test.

Reviewed By: arthaud

Differential Revision: D46811051

fbshipit-source-id: f38e06d135d0b662abc93712e6e5f5bc083aff08